rgwam: reorganize code
