+++ src/pybind/mgr/rgw/__init__.py
@@ -2,7 +2,3 @@ try:
     from .module import Module
 except ImportError:
     pass
-
-import logging
-
-log = logging.getLogger(__name__)
