+++ src/python-common/ceph/rgw/__init__.py
@@ -1,0 +1,3 @@
+import logging
+
+log = logging.getLogger(__name__)
