+++ src/python-common/ceph/rgw/diff.py
