+++ src/python-common/ceph/rgw/rgwam_core.py
@@ -1,4 +1,3 @@
-#!@Python3_EXECUTABLE@
 # -*- mode:python -*-
 # vim: ts=4 sw=4 smarttab expandtab
 #
