+++ src/python-common/ceph/rgw/types.py
